third_party/WebKit/LayoutTests/crypto/subtle/hkdf/deriveBits-expected.txt

Test that HKDF does not support methods it should not support.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Derive 0 bits from the HKDF key
PASS derivedBits.byteLength is 0
Derive 4 bits from the HKDF key
PASS derivedBits.byteLength is 1
PASS derivedBits.getUint8(0) is 0x80
PASS successfullyParsed is true

TEST COMPLETE
